优化客户信息类，采用BigDecial计算积分和消费额

+++ src/main/java/com/coding/sales/customer/CustomerFactory.java
@@ -1,6 +1,8 @@
 package com.coding.sales.customer;
 
+import java.math.BigDecimal;
+
 public interface CustomerFactory {
-	public Customer makeCustomer(String name, String cardNumber, int points);
+	public Customer makeCustomer(String name, String cardNumber, BigDecimal points);
 	public Customer upgradeCustomer(Customer customer);
 }
