+++ src/main/java/com/coding/sales/customer/CustomerRating.java
@@ -1,6 +1,8 @@
 package com.coding.sales.customer;
 
+import java.math.BigDecimal;
+
 public interface CustomerRating {
-	public void earnedPoints(int amountMoney);
+	public void earnedPoints(BigDecimal amountMoney);
 	public void makeSureCustomerShouldUpgrade(int upgradePoints);
 }
